+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 
+- Allow switching off the built-in Passport error handling with `Visa::disablePassportErrorHandling()`
+- Replace `AccessTokenController` to use the replacement `HandlesOAuthErrors` trait
+- Replace `ApproveAuthorizationController` to use the replacement `HandlesOAuthErrors` trait
+- Replace `AuthorizationController` to use the replacement `HandlesOAuthErrors` trait
+- Add a replacement `HandlesOAuthErrors` trait that can toggle error handling
 - Add `CheckFirstPartyClient` middleware to prevent non-first-party clients from accessing assigned endpoints
 - Allow switching to UUIDs instead of random strings for client IDs with `Visa::enableClientUUIDs()`
 - Replace `Client` class to set `$incrementing` to `false` in support of string or UUID client IDs